pyproject.toml
[project]
name = "rest-api"
version = "0.1.0"
description = ""
authors = [
    {name = "gabriel-garciae",email = "eng.gabrielgevangelista@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "joblib (>=1.5.1,<2.0.0)",
    "numpy (>=2.3.2,<3.0.0)",
    "scikit-learn (>=1.7.1,<2.0.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "psycopg2-binary (>=2.9.9,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
